qwerty

+++ problems/problem_1.cpp
@@ -1,7 +1,7 @@
-void problemSolution1(double consumed_water) {
+void problemSolution1(float consumed_water) {
     float cost;
     // write your code here
-    cost = 13;
+    cost = 13.0;
     if(consumed_water > 0){
         cost += min(30.0, consumed_water) * 0.4;
         consumed_water -= 30;
